modificando de inteiro pra ponto boiante

+++ src/simulacao.h
@@ -23,7 +23,7 @@ typedef struct
   int *frequencias; // Lista que contém quantas vezes o objeto foi acessado.
   float *hits; // Armazena o número de hits / camada.
   float *misses; // Armazena o número de misses / camada.
-  int reqTotal; // Armazena o total de requests feitos no total.
+  float reqTotal; // Armazena o total de requests feitos no total.
   float tempoTotal; // Armazena o tempo de resposta total gasto nas camadas.
 } Simulacao;
